GrowBox_Arduino
This file has the added code for the Ultrasonic sensor that i didnt use in the end because i didnt think i needed it. 

I have added in the blynk and the ESP-01 code also to show this is what it would do if i managed to get it working.  
Some of the code is commented out because i either couldnt get it working or i had an MCU problem with it.